refactor: move sequence model to base models
