+++ cyberdrop_dl/models/base_models.py
@@ -1,13 +1,25 @@
 """Pydantic models"""
 
-from collections.abc import Mapping
+from collections.abc import Iterator, Mapping, Sequence
 from pathlib import Path
+from typing import TypeVar
 
 import yarl
-from pydantic import AnyUrl, BaseModel, ConfigDict, Secret, SerializationInfo, model_serializer, model_validator
+from pydantic import (
+    AnyUrl,
+    BaseModel,
+    ConfigDict,
+    RootModel,
+    Secret,
+    SerializationInfo,
+    model_serializer,
+    model_validator,
+)
 
 from cyberdrop_dl.models.types import HttpURL
 from cyberdrop_dl.models.validators import to_apprise_url_dict
+
+_ModelT = TypeVar("_ModelT", bound=BaseModel)
 
 
 class AliasModel(BaseModel):
@@ -51,3 +63,17 @@ class PathAliasModel(AliasModel):
                 setattr(self, name, self.replace_config_and_resolve(value, config_name))
             elif isinstance(value, PathAliasModel):
                 value.resolve_paths(config_name)
+
+
+class SequenceModel(RootModel[list[_ModelT]], Sequence[_ModelT]):
+    def __len__(self) -> int:
+        return len(self.root)
+
+    def __iter__(self) -> Iterator[_ModelT]:
+        yield from self.root
+
+    def __getitem__(self, index: int) -> _ModelT:
+        return self.root[index]
+
+    def __bool__(self) -> bool:
+        return bool(len(self))
